+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotTemplateParameter.java
@@ -1,20 +1,16 @@
 // This is a generated file. Not intended for manual editing.
 package dev.xeonkryptos.xeonrobotframeworkplugin.psi.element;
 
-import java.util.List;
-import org.jetbrains.annotations.*;
-import com.intellij.psi.PsiElement;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public interface RobotTemplateParameter extends RobotArgument, RobotStatement {
 
   @Nullable
-  RobotTemplateParameterArgument getTemplateParameterArgument();
+  RobotTemplateArgument getTemplateArgument();
 
   @NotNull
   RobotTemplateParameterId getTemplateParameterId();
-
-  @Nullable
-  RobotVariable getVariable();
 
   @NotNull String getParameterName();