+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/impl/RobotTemplateParameterImpl.java
@@ -1,15 +1,16 @@
 // This is a generated file. Not intended for manual editing.
 package dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.impl;
 
-import java.util.List;
-import org.jetbrains.annotations.*;
 import com.intellij.lang.ASTNode;
-import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
-import static dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotTypes.*;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.*;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotPsiImplUtil;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTemplateArgument;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTemplateParameter;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotTemplateParameterId;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVisitor;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class RobotTemplateParameterImpl extends RobotPsiElementBase implements RobotTemplateParameter {
 
@@ -29,20 +30,14 @@ public class RobotTemplateParameterImpl extends RobotPsiElementBase implements R
 
   @Override
   @Nullable
-  public RobotTemplateParameterArgument getTemplateParameterArgument() {
-    return PsiTreeUtil.getChildOfType(this, RobotTemplateParameterArgument.class);
+  public RobotTemplateArgument getTemplateArgument() {
+    return PsiTreeUtil.getChildOfType(this, RobotTemplateArgument.class);
   }
 
   @Override
   @NotNull
   public RobotTemplateParameterId getTemplateParameterId() {
     return notNullChild(PsiTreeUtil.getChildOfType(this, RobotTemplateParameterId.class));
-  }
-
-  @Override
-  @Nullable
-  public RobotVariable getVariable() {
-    return PsiTreeUtil.getChildOfType(this, RobotVariable.class);
   }
 
   @Override